Update patch set 3
Patch Set 3:

Hmmm, a quick (non-owner) check to see if we have a non-trivial merge is for bust; because of the original import was in itself a port to Android and dropped a lot of content prior to showing its face to our repository. Lorenzo, you may have to be the owner of the merge to 2.6.0 (the sound of someone being eaten by the wolves in the background ;-/)

None of the fixes I have in this CL are present in the latest code.

Patch-set: 3