Update patch set 3
Patch Set 3:

Lorenzo, we will see you on Thursday with the rest of the system team. The systems group opinion is also to remove it unless someone is passionate about maintaining it. @jpa believes it will be a simple port from upstream as there were some patches applied to it naming android; he has a few tools that can aid transliterating the 2.6.0 source ball directly into external/. Lets discuss the disposition of netperf in the meeting as part of your agenda?

Patch-set: 3