Update patch set 3
Patch Set 3: Code-Review+2

I think we should merge this change.

I agree that if this turns out to be a real maintenance burden, we should either delete it or update it to upstream. For now, given that we already have a fix in hand, the best tradeoff seems to be to fix it.

That way, those that are using it can keep using it, without having to spend cycles on merging upstream, and we achieve the goal of building with -Wall -Werror. Yes, we won't have a fully up-to-date version, and updating to upstream will cause lots of conflicts, but I suspect that's no different to where we are today.

If this keeps causing maintenance burden, we can always delete it or update it later.

Patch-set: 3
Reviewer: Lorenzo Colitti <1000835@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2